refactor(desktop): combine remote backend connection into one module
backendStatus, backendRedirects, and backendHopRequest each exposed part of
one job, so connecting to a remote backend required understanding the hop
transport, the redirect policy, and the probe sequence separately.

remoteBackends exposes connectRemoteBackend and the result it returns. The
transport, redirect policy, probes, and retry budget are now implementation
details, and the net.request transport is the default rather than a caller
argument.
